A more efficient fix for aspd buffs.
This uses the same approach as the v1 fix, but checks a variable set directly on the unit itself instead of checking a global table.

Manually set the can_use_aspd variable to true in each unit section in Player:init, where the unit has a timed action with tag 'shoot' or 'attack'.